requirements.txt
Flask~=3.1.1
Flask-SQLAlchemy
Flask-Migrate~=4.1.0
python-dotenv~=1.1.0
pyodbc
SQLAlchemy~=2.0.41
alembic~=1.15.2